fix: partially fix version comparison and splitting of number-like segments
The previous regex was incorrect; formatting that `|` is a special char in regex, which meant the splitting on |, : and - were not functional. Based on the comments, the intent was for all 3, and a period, to also act as a version splitter. This caused the logic to compare some elements as integers and some as strings which broke the compareTo contract. This is still problematic as the contract breaks as soon as there is an ASCII char < `0` which is not defined as a splitter.